pyrequirement.txt
paho-mqtt
pymongo[srv]
pyshark
configparser
datetime
requests
